Directory refactoring completion. Model settings file format changed to YAML.
