+++ src/__init__.py
@@ -1,0 +1,2 @@
+"src folder"
+
