+++ src/frontend/__init__.py
@@ -1,0 +1,2 @@
+"frontend folder"
+
