+++ src/log/__init__.py
@@ -1,0 +1,2 @@
+"log folder"
+
